fix #18 ①JSONParseとXMLParseをextensionファイルとして、別ファイルに定義 ②UISwitchでJSONとXMLのパース切り替え実装（Bool初期値はJSONパース）
